Init Collab

+++ FinancialTransaction+CoreDataProperties.swift
@@ -2,7 +2,7 @@
 //  FinancialTransaction+CoreDataProperties.swift
 //  FirstChallange
 //
-//  Created by Syauqi Ikhlasun Nadhif on 20/03/25.
+//  Created by Syauqi Ikhlasun Nadhif on 21/03/25.
 //
 //
 
@@ -16,7 +16,7 @@ extension FinancialTransaction {
         return NSFetchRequest<FinancialTransaction>(entityName: "FinancialTransaction")
     }
 
-    @NSManaged public var amount: Double
+    @NSManaged public var amount: Int64
     @NSManaged public var category: String?
     @NSManaged public var date: Date?
     @NSManaged public var id: UUID?
